% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-check.R
\name{data_check}
\alias{data_check}
\title{Data checking}
\usage{
data_check(x)
}
\arguments{
\item{x}{a \code{matrix} or \code{data.frame}.}
}
\value{
a \code{matrix} containing the cleaned data after running the 
    necessary checks.
}
\description{
Checks the integrity of the inputed data before computing the 
Euclidian Minimum Spanning Tree (EMST)
}
\details{
\code{data_check} is called from inside 
    \code{\link[emstreeR]{ComputeMST}} before the computation begins. First, 
    it evaluates the object format. Afterwards, it checks whether the 
    inputed data has at least two columns and tries to coerce all columns 
    into numeric, beyond removing all rows containing \code{NA}'s entries.
}
