% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-MST.R
\docType{data}
\name{emstreeR-ggproto}
\alias{emstreeR-ggproto}
\alias{StatMST}
\title{'ggproto' objects}
\description{
Internal use only definitions of 'ggproto' objects needed for
  geoms and stats.
}
\keyword{datasets}
\keyword{internal}
