% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mlpack_mst}
\alias{mlpack_mst}
\title{\code{mlpack}'s Euclidean Minimum Spanning Tree}
\usage{
mlpack_mst(data)
}
\arguments{
\item{data}{a \code{matrix}.}
}
\value{
a \code{matrix} containing each pair of start and end points
on its columns, and the distance between these points in order to 
produce the Minimum Spanning Tree.
}
\description{
Fast computes an EMST using the \code{mlpack C++} library.
}
